requirements.txt
# ============================================
# Core dependencies
# ============================================
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.2.0

# ============================================
# Deep Learning
# ============================================
torch>=2.0.0
transformers>=4.35.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
sentence-transformers>=2.2.0
datasets>=2.14.0

# ============================================
# NLP
# ============================================
tokenizers>=0.14.0
sacrebleu>=2.3.0
rouge-score>=0.1.2
nltk>=3.8.0

# ============================================
# Graph Neural Networks (Chapter 3)
# ============================================
torch-geometric>=2.3.0
networkx>=3.0

# ============================================
# Sequence Labeling (Chapter 4)
# ============================================
torchcrf>=1.1.0
seqeval>=1.2.2

# ============================================
# Visualization
# ============================================
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.64.0

# ============================================
# Geocoding (Chapter 2)
# ============================================
geopy>=2.3.0
requests>=2.28.0

# ============================================
# Utilities
# ============================================
pyyaml>=6.0
jsonlines>=3.1.0
